+++ src/net/libwebsocketsIO.h
@@ -44,7 +44,7 @@ protected:
     size_t getOutputBufferLength();
     void resetOutputBuffer();
     
-    virtual bool wsSendMessage(char *msg, uint64_t len);
+    virtual bool wsSendMessage(char *msg, size_t len);
     virtual void wsDisconnect(bool immediate);
     virtual bool wsIsConnected();
