+++ src/net/libwsIO.cpp
@@ -167,7 +167,7 @@ void LibwsClient::websockMsgCb(ws_t ws, char *msg, uint64_t len, int binary, voi
     }, self->appCtx);
 }
                          
-bool LibwsClient::wsSendMessage(char *msg, uint64_t len)
+bool LibwsClient::wsSendMessage(char *msg, size_t len)
 {
     assert (mWebSocket);
